continue (#15)
- [x] make docker file shorter than 4000 characters...
- [x] fix issue with missing date
- [x] retry on count timeout
- [x] only push images if new
- [x] keep hash and count in data
- [x] improve chunking
- [x] stay consistent on insert failure
- [x] ~~be smarter about caching~~ duplicate payload -- doc contains
values, snippets contain lists
- [x] move spacy to scattermind worker
- [x] test out the xx model
- [x] ~~add main ids to the filter below a certain number of expected
results~~
- [x] ~~parse main language from url if possible~~
- [x] lower relevance of articles with unknown dates
- [x] prevent 1970 dates
- [x] make a filter respect relevance of field values
- [x] add relevance to field values (language prob, country count, etc.)
- [x] scale up qdrant
- [x] unify chunking code
- [x] return ordered list of countries / languages / etc.
- [x] ~~infer date from url as well~~
- [x] infer country from url as well
- [x] order by date and most relevant country (and make it user
decidable)
- [x] add field to indicate invalid date (updated contains the update
date. if date and updated are the same it is an invalid date)
- [x] use first few words if title is missing
- [x] snippify endpoint and improve snippets
- [x] move special cases from notebook to api
- [x] avoid non-alpha starts of snippets
- [x] treat NUL as no country or warn (NUL can appear in unexpected
places. OpenCage returns NUL for Kosovo)
- [x] fix qdrant dashboard
- [x] make hit limit an argument
- [x] add date module
- [x] add OpenAPI (for insights)

+++ sh/dockerpush.sh
@@ -6,7 +6,7 @@ cd -- "$( dirname -- "${BASH_SOURCE[0]}" )/../" &> /dev/null
 
 IMAGE_TAG="${IMAGE_TAG:-$(make -s name)}"
 IMAGE_BASE="nlpapi"
-DOCKER_LOGIN_SERVER="acclabdocker.azurecr.io"
+DOCKER_LOGIN_SERVER="acclabdocker.azurecr.io/"
 
 DEVMODE_CONF_FILE="deploy/devmode.conf"
 REDIS_VERSION_FILE="deploy/redis.version"
@@ -26,17 +26,17 @@ fi
 
 dpush() {
     IMAGE="${IMAGE_BASE}-$1:$2"
-    URL="${DOCKER_LOGIN_SERVER}/${IMAGE}"
-    # if ! docker manifest inspect "${URL}" &> /dev/null ; then
-    echo "pushing ${IMAGE} to ${URL}"
+    URL="${DOCKER_LOGIN_SERVER}${IMAGE}"
+    if ! docker pull "${URL}" &> /dev/null ; then
+        echo "pushing ${IMAGE} to ${URL}"
 
-    docker tag "${IMAGE}" "${URL}"
-    docker push "${URL}"
-    docker rmi "${URL}"
-    # else
-    #     echo "${URL} already exists"
-    # fi
-    sleep 1
+        docker tag "${IMAGE}" "${URL}"
+        docker push "${URL}"
+        docker rmi "${URL}"
+        sleep 1
+    else
+        echo "${URL} already exists"
+    fi
 }
 
 dpush "worker" "${IMAGE_TAG}"
@@ -51,13 +51,13 @@ QDRANT_BASE="qdrant/qdrant:v1.8.0"
 docker pull --platform linux/amd64 "${QDRANT_BASE}"
 
 IMAGE_QDRANT="${IMAGE_BASE}-qdrant:${QDRANT_DOCKER_VERSION}"
-URL_QDRANT="${DOCKER_LOGIN_SERVER}/${IMAGE_QDRANT}"
-# if ! docker manifest inspect "${URL_QDRANT}" &> /dev/null ; then
-echo "pushing ${QDRANT_BASE} to ${URL_QDRANT}"
+URL_QDRANT="${DOCKER_LOGIN_SERVER}${IMAGE_QDRANT}"
+if ! docker pull "${URL_QDRANT}" &> /dev/null ; then
+    echo "pushing ${QDRANT_BASE} to ${URL_QDRANT}"
 
-docker tag "${QDRANT_BASE}" "${URL_QDRANT}"
-docker push "${URL_QDRANT}"
-docker rmi "${URL_QDRANT}"
-# else
-#     echo "${URL_QDRANT} already exists"
-# fi
+    docker tag "${QDRANT_BASE}" "${URL_QDRANT}"
+    docker push "${URL_QDRANT}"
+    docker rmi "${URL_QDRANT}"
+else
+    echo "${URL_QDRANT} already exists"
+fi
